Initial test refactoring according the new authentication and authorizazione impl

+++ auth-service/src/main/scala/io/sherpair/w4s/auth/repository/RepositoryOps.scala
@@ -5,7 +5,7 @@ import fs2.Stream
 trait RepositoryOps[F[_], K, R] {
 
   /* Test-only */
-  def count: F[Int]
+  def count: F[Long]
 
   def delete(id: K): F[Int]
